+++ src/frontend/shared/components/NavigationItems.tsx
@@ -1,13 +1,23 @@
 import * as React from 'react';
 import {Link} from "react-router-dom";
-import {IUserProps} from "../interfaces/IUserProps";
 import {connect} from "react-redux";
+import {loadDisplayedFriends, loadDisplayedUser} from "../../settings/actions";
+import {IUser} from "../../../shared/ModelInterfaces";
+import {getManyUsersInfo} from "../globleFunctions";
 
 interface INavigationItemsProps {
-    userInfo: any
+    userInfo: IUser,
+    loadDisplayedUser: any,
+    loadDisplayedFriends: any
 }
 
 class NavigationItems extends React.Component<INavigationItemsProps, {}> {
+
+    loadUserProfile = async () => {
+        this.props.loadDisplayedUser(this.props.userInfo);
+        const friendList = await getManyUsersInfo(this.props.userInfo.friendUsernames);
+        this.props.loadDisplayedFriends(friendList);
+    };
 
     render() {
         return (
@@ -16,7 +26,7 @@ class NavigationItems extends React.Component<INavigationItemsProps, {}> {
                     width="100" height="100"/><p className="username-display">{"   " + this.props.userInfo.fullname}</p></a>
             <Link to="/" className="glyphicon glyphicon-home">Home</Link>
             <Link to="/chatRoom" className="glyphicon glyphicon-envelope">Chats</Link>
-            <Link to="/settings" className="glyphicon glyphicon-cog">Profile</Link>
+            <Link to="/settings" className="glyphicon glyphicon-cog" onClick={this.loadUserProfile}>Profile</Link>
             <a href="/api/v1/users/logout" className="glyphicon glyphicon-log-out">Logout</a>
         </div>
         );
@@ -29,4 +39,4 @@ const mapStateToProps = (state: { userInfo: any }) => {
     };
 };
 
-export default connect(mapStateToProps)(NavigationItems);
+export default connect(mapStateToProps, {loadDisplayedUser, loadDisplayedFriends})(NavigationItems);